feat: rework architecture
