+++ python/beeai_framework/adapters/acp/serve/__init__.py
@@ -12,6 +12,3 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from beeai_framework.adapters.acp.serve.server import AcpAgentServer
-
-__all__ = ["AcpAgentServer"]
